Split ResourceManager implementations into own files + fix tests
